Refactored PGClient API and internals.
No need to use an explicit latch, Bootstrap has this functionality
built in.

InterruptedException should be propagated to the caller.

Give explicit names to the pipeline stages. I'm using a "PG" prefix here
as we may want to allow users to add pipeline stages in future and it
may be nice to keep this as a reserved prefix to tell user stages from
ours.