Refactored FeBeMessageType
Improvements to FeBeMessageType, to be self-contained. Returned to
initial formatting of the enum constructor for easier human reading.
Avoid using external constants for the values. Provide cached byte type
rather than Character, and provide helper method hasType() to avoid
comparing to a hardwired constant (NONE).

Fixed a bug in the 4-args constructor of the enum, length field was
ignored.

Deleted MessageId, no longer necessary. There are some messages common
to both fe and be and that wasn't reflected in MessageId.