Commented out tests that depend on a certain configuration of external systems
Added a TODO note to create a test environment that does not depend on external
systems or, at least, on their current configuration.  As noted in the
comments, https://github.com/airlift/testing-postgresql-server might be a good
alternative.

In the meantime, tests need to be deactivated or builds break unless a external
PostgreSQL database is specifically configured for the tests, which is IMHO not
a good thing.