Fixed possible publication issue in PGEncoding
PGEncoding has a map that is shared by all enum instances. Reads need to
be lock free, as they may be very frequent. Writes are very infrequent.
Writes where protected by double-checking and a lock, reads with no
contention.

But reads may have read a partially put object. This patch adds a
volatile field as a memory barrier, so that put is guaranteed to happen
before the writing process returns, this allowing safe publication.

Also two encodings are initialized and cached by default: US-ASCII (used
during startup phase) and UTF-8 (which should be the encoding to use
almost all the time).