Significant rewrite of the Phoebe project
Complete new implementation of the Message strategy. This patch changes
the notion of Messages from being regular Java POJOs to one where
Messages are simply views of a given, immutable ByteBuf.

Netty's ByteBuf becomes the centerpiece of the Message. The decoding
process is as simple as storing a reference to the ByteBuf, and provide
getters that are just read*() calls with an offset over the ByteBuf. The
encoding process is more or less the same (allocate the ByteBuf and
write to it) but happens at construction time. Convenience methods are
provided in the Messages and in MessageEncoders to easily create new
Messages.

With this change all the encoding and decoding business logic is now
bundled with the message itself, and not spread out through three files
(the message, the encoder and the decoder). It should also improve heap
allocation and GC as messages are now tiny (only contain a reference to
the ByteBuf and the Charset), compared to allocating different objects
with class attributes for every message field. The ByteBuf is allocated
off heap and pooled, and sizes are optimized to power-of-two to help
Netty allocate them more efficiently.

Messages are now encoded/decoded in two parts: the header and the body.
They are then "merged" together by a CompositeByteBuf. This approach has
some benefits, like: separation of concerns, not requiring memory
for the header in the decoding process (only encoding) and re-using
headers for messages with constant headers (not included in this patch,
TODO).

All in all this patch is about simplification and --hopefully-- improved
performance and less GC. This patch introduces a significant reduction
both in terms of lines of code and files. Creating new messages is
probably simpler and both the encoding and decoding logics are now
provided, which means that more functionality is provided than before
(although this contributes only to help creating not only PostgreSQL
clients, but also servers).