ByteBufAllocatorUtil: centralize ByteBuf alloc
To create the Message bodies and headers, we explicitly allocate the
ByteBufs. But there was no "central" place for doing so. This patch
creates class ByteBufAllocatorUtil with static methods for it. It
provides a method for allocating ByteBuf which might contain strings
encoded into them, and another one for ByteBuf which will not contain
strings. The latter ones result in smaller capacity ByteBuf.