Gracefully terminate PGClient sessions
The PGClient and PGSession still need a lot of love, but at least basic
support has been added to gracefully terminate programs. A close()
method has been added to PGSession, to release any internal resources
acquired, and a disconnect() method has been added to PGClient, to
terminate the Netty daemons, previously calling PGClient.close(). So it
suffices to disconnect() from the client to end the program.