First version of a Reactive Client API
This commit introduces a new alternative Client API based on the
Reactive Streams (http://www.reactive-streams.org/) initiative.
Its implementation is based on the RxJava
(https://github.com/ReactiveX/RxJava) libraries.

The Client API is defined in module phoebe-client-rs, in the
interface PostgresClient, which exposes the
connection(s) via ReactiveStream's Publisher interface. The
implementation is present in module phoebe-client-rs-rxjava, in the
class RxPostgresClient, which also exposes RxJava's Observable methods
for convenience.

This is WIP, but at some point should allow to delete the current
phoebe-client-api module.