Defensive programming to help phoebe programs exit
Users of Phoebe should be always calling client.close(). Before this
change, failing to do so would result in the JVM not exiting, because of
the EventLoopGroup threads being non-daemon.

This patch changes the reads created by Netty's EventLoopGroup to be
daemons and adding a JVM shutdownHook to call close() even if the user
doesn't call close(), in an attempt to cleanly tear down connections.

Now the program always exits quickly and tears down connections even if
the user is not calling .close().